Proper parallelism support.
1. Create num_threads/2 read strands (and write strands)

2. Cycle through the strands each time a new session is created.

This allows long operations to execute in parallel.  Any futher concurrency
(for application logic) will utilize coroutines.  Thus, a server application
will have exactly the same number of threads as cores, regardless of
the level of logical concurrency.  All user code responding to an incoming
message will be running on a single strand.